pyproject.toml
[project]
name = "mcp-url2markdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.6.0",
    "fastmcp>=2.2.2",
]
